+++ Reika/SatisfactoryPlanner/Data/Objects/ResourceSupplies/TruckStation.java
@@ -17,7 +17,7 @@ public class TruckStation extends LogisticSupply<TruckStation, Item> {
 
 	private TruckStation(JSONObject obj) {
 		this((Item)Database.lookupItem(obj.getString("item")));
-		this.setAmount(obj.getInt("amount"));
+		this.setAmount(obj.getDouble("amount"));
 	}
 
 	@Override
