+++ Reika/SatisfactoryPlanner/FactoryListener.java
@@ -23,7 +23,7 @@ public interface FactoryListener extends Comparable<FactoryListener> {
 	public void onRemoveRecipe(Recipe r);
 	public void onRemoveRecipes(Collection<Recipe> c);
 	public void onSetCount(Recipe r, double count);
-	public void onSetCount(Generator g, Fuel fuel, int old, int count);
+	public void onSetCount(Generator g, Fuel fuel, double old, double count);
 	public void onAddProduct(Consumable c);
 	public void onRemoveProduct(Consumable c);
 	public void onRemoveProducts(Collection<Consumable> c);
